# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WatchIT Tower Monitoramento — API
  version: "1.0.0"
  description: |
    API REST do WatchIT Tower Monitoramento — dashboard local de monitoramento de infraestrutura de TI.

    ## Autenticação

    A maioria dos endpoints exige uma sessão NextAuth ativa (cookie de sessão JWT).
    Faça login em `POST /api/auth/callback/credentials` e inclua o cookie `next-auth.session-token`
    nas requisições subsequentes.

    ### Exceções (sem autenticação de sessão)
    - `GET /api/health` — público, sem autenticação
    - `POST /api/links/{id}/up` e `POST /api/links/{id}/down` — autenticados por HMAC-SHA256 via header `x-webhook-token` ou query string `?token=`

    ## Rate Limiting
    O endpoint de login (`POST /api/auth/callback/credentials`) tem rate limit de
    **10 tentativas por IP em 15 minutos**.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Cookie de sessão JWT gerado pelo NextAuth após login bem-sucedido.

    webhookToken:
      type: apiKey
      in: header
      name: x-webhook-token
      description: |
        Token HMAC-SHA256 gerado com `WEBHOOK_SECRET + linkId`.
        Pode ser enviado como header `x-webhook-token` ou query string `?token=`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
              example: "Unauthorized"
            - type: object
              description: Objeto de erros de validação Zod (fieldErrors + formErrors).
      required: [error]
...
security:
  - sessionCookie: []

# ─── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /api/auth/callback/credentials:
    post:
      tags: [Autenticação]
      summary: Login com usuário e senha
      description: |
        Endpoint gerenciado pelo NextAuth. Autentica o usuário e retorna um cookie de sessão JWT.

        **Rate limit:** 10 tentativas por IP em 15 minutos. Retorna `429` quando o limite é excedido.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  format: password
                  example: "senha-segura-123"
      responses:
        "200":
          description: Login bem-sucedido. O cookie `next-auth.session-token` é definido na resposta.
        "401":
          description: Credenciais inválidas.
        "429":
          description: Muitas tentativas. Aguarde 15 minutos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Muitas tentativas de login. Aguarde 15 minutos."

  # ── Sistema ───────────────────────────────────────────────────────────────
...